add git unit testing
