+++ crates/playground/js/App.tsx
@@ -1,26 +1,7 @@
-import { useEffect, useState } from "react";
-import * as Comlink from "comlink";
-import { Api } from "./wasm-worker";
-
-let start = async (setter: (message: String) => void) => {
-  const handlers = await Comlink.wrap<Api>(
-    new Worker(new URL("./wasm-worker.ts", import.meta.url), {
-      type: "module",
-    })
-  ).handlers;
-  let test = await handlers.greet("ink Playground");
-  setter(test);
-};
-
 const App = () => {
-  const [message, setMessage] = useState<String>("Check");
-  useEffect(() => {
-    start(setMessage);
-  });
-
   return (
     <>
-      <h1>{message}</h1>
+      <h1>Welcome to ink Playground!</h1>
     </>
   );
 };
