+++ crates/playground/js/__tests__/App.test.tsx
@@ -1,8 +1,8 @@
 import { render, screen } from "@testing-library/react";
 import App from "../App";
 
-test("renders Basic TypeScript App", () => {
+test("Contains Welcome Message", () => {
   render(<App />);
-  const linkElement = screen.getByText(/Basic TypeScript App/i);
+  const linkElement = screen.getByText(/Welcome to ink Playground!/i);
   expect(linkElement).toBeInTheDocument();
 });
