+++ crates/playground/package.json
@@ -8,7 +8,7 @@
   "scripts": {
     "build": "rimraf dist pkg && webpack --config webpack.config.js --mode production",
     "start": "rimraf dist pkg && webpack serve --host 0.0.0.0",
-    "test": "wasm-pack test --headless --chrome --safari --firefox & jest"
+    "test": "wasm-pack test --headless --chrome & jest"
   },
   "devDependencies": {
     "@testing-library/jest-dom": "^5.14.1",
